More accurate initial position for pid_main_ma
pid_main_ma.py: Previously, the pid controller set the center of the
image frame as the assumed starting point of the colored node. This was
updated so that the program would wait for a colored node to be detected
in the frame first, then select that as the initial position.

Autotuning the PID gains using MATLAB was attempted. It was attempted to
linearize the system using the MATLAB system identification toolbox, for
a state space model with 15 states. However, the system with the auto-
tuned gains applied was not observed to be more stable than when using
the manually tuned gains.